requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
absl-py==2.1.0
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.68.1
huggingface-hub==0.27.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.14.3
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.29.2
psutil==6.1.1
ptflops==0.7.4
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.6.1
scikit-video==1.1.11
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.20.0
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.13
tokenizers==0.21.0
torch==2.2.1+cu121
torch-fidelity==0.3.0
torchaudio==2.2.1+cu121
torchmetrics==1.7.1
torchvision==0.17.1+cu121
tqdm==4.67.1
transformers==4.47.1
triton==2.2.0
typed-argument-parser==1.10.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wandb==0.19.6
Werkzeug==3.1.3
xformers==0.0.25
zipp==3.21.0
